Functions Update
After reviewing the above backup.sh script, you will notice the following changes to the code:
we have defined a new function called total_files. The function utilized the find and wc commands to determine the number of files located within a directory supplied to it during the function call
we have defined a new function called total_directories. Same as the above total_files function it utilized the find and wc commands however it reports a number of directories within a directory supplied to it during the function call.