Conditional Statements Update
Lines 15 - 21 are used to define two new functions returning a total number of files and directories included within the resulting compressed backup file. After the backup Line 23 is executed, on Lines 25 - 29 we declare new variables to hold the total number of source and destination files and directories.

The variables concerning backed up files are later used on Lines 36 - 42 as part of our new conditional if/then/else statement returning a message about the successful backup on Lines 37 - 39only if the total number of both, source and destination backup files is equal as stated on Line 36.