Positional Parameters Update
The above backup.sh script update introduces few new bash scripting techniques but rest for the code between Lines 5 - 13 should be by now self-explanatory. Line 5 is using a -z bash option in combination with conditional if statement to check whether positional parameter $1 contains any value. -z simply returns true if the length of the string which in our case is variable $1 is zero. If this is the case, we set $user variable to a current user's name.